requirements.txt
openai>=1.2,<2
fastapi
uvicorn
python-dotenv
jinja2
aiofiles
python-multipart